Moving IMonoalphabeticSubstitutionCipher to new project.
