+++ SimpleCryptoLib/Ciphers/Monoalphabetic Substitution Cipher/IMonoalphabeticSubstitutionCipher.cs
@@ -1,0 +1,7 @@
+using SimpleCryptoLib.Ciphers.Common;
+
+namespace SimpleCryptoLib.Ciphers.Monoalphabetic_Substitution_Cipher;
+
+public interface IMonoalphabeticSubstitutionCipher : ICommonCipher<MonoalphabeticSubstitutionKey>
+{
+}
